Inclusão da recuperação de senha do cliente

+++ database/migrations/2017_10_20_145121_create_order_items_table.php
@@ -17,10 +17,10 @@ class CreateOrderItemsTable extends Migration
             $table->increments('id');
             $table->integer('order_id')->nullable();
             $table->integer('product_id')->nullable();
-            $table->decimal('preco', 5, 2)->nullable()->unsigned();
-            $table->decimal('qtde', 5, 2)->nullable()->unsigned();
-            $table->decimal('total', 5, 2)->nullable()->unsigned();
-            $table->decimal('pontos', 5, 2)->nullable()->unsigned();
+            $table->decimal('preco', 13, 2)->nullable()->unsigned();
+            $table->integer('qtde')->nullable()->unsigned();
+            $table->decimal('total', 13, 2)->nullable()->unsigned();
+            $table->decimal('pontos', 13, 2)->nullable()->unsigned();
             $table->timestamps();
             //$table->foreign('order_id')->references('id')->on('orders');
             //$table->foreign('product_id')->references('id')->on('products');
